feat(data): recency-decay weighting for pair and counter counts
Balance patches invalidate old comps overnight; a hard 60-day window kept
them at full weight until they vanished. Matches now decay with a 21-day
half-life inside an 84-day guard, so the tables track the current patch
automatically. Counts become fractional; the schema drops the int
constraint (old clients reject the new file and degrade to zero terms
until their cached build refreshes).

+++ src/web/lib/data/pairs-schema.ts
@@ -9,9 +9,10 @@ import { z } from "zod";
 
 export const PAIRS_SCHEMA_VERSION = 1;
 
+/** Fractional: matches are recency-weighted (half-life decay), not counted. */
 const countSchema = z.object({
-  matches: z.number().int().nonnegative(),
-  wins: z.number().int().nonnegative(),
+  matches: z.number().nonnegative(),
+  wins: z.number().nonnegative(),
 });
 
 export const pairsTableSchema = z.object({
